small fix
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@3495 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VBoxBFE/EmulCpp.cpp
@@ -31,7 +31,7 @@
 
 /** libstdc++ new emulator. */
 void*
-operator new(unsigned int size)
+operator new(size_t size)
 {
     return malloc(size);
 }
